tools/requirements.txt
# Python dependencies (for scripts/yaml2rdf)
click
packaging
pyyaml
rdflib
wget
gitpython
